Fill dropdown menus and UI upgrades

+++ CRUD_Employees/Controllers/EmployeesController.cs
@@ -10,6 +10,8 @@ using CRUD_Employees.Models;
 using Microsoft.AspNetCore.Hosting;
 using Microsoft.AspNetCore.Http;
 using System.IO;
+using CRUD_Employees.Helpers;
+using CRUD_Employees.Models.Enum;
 
 namespace CRUD_Employees.Controllers
 {
@@ -50,6 +52,10 @@ namespace CRUD_Employees.Controllers
         // GET: Employees/Create
         public IActionResult Create()
         {
+            ViewBag.Genders = EnumHelper.GetEnumSelectList<Gender>();
+            ViewBag.ContractTypes = EnumHelper.GetEnumSelectList<ContractType>();
+            ViewBag.Departments = EnumHelper.GetEnumSelectList<Department>();
+
             return View();
         }
 
@@ -64,7 +70,6 @@ namespace CRUD_Employees.Controllers
             {
                 if (imageFile != null)
                 {
-                    // Process the uploaded file
                     string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, "images");
                     string uniqueFileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
                     string filePath = Path.Combine(uploadsFolder, uniqueFileName);
@@ -74,14 +79,23 @@ namespace CRUD_Employees.Controllers
                         await imageFile.CopyToAsync(fileStream);
                     }
 
-                    // Store the relative path in the database
                     employee.Image = "images/" + uniqueFileName;
+                }
+
+                if (employee.ContractType == ContractType.Permanent)
+                {
+                    employee.ContractDue = null;
                 }
 
                 _context.Add(employee);
                 await _context.SaveChangesAsync();
                 return RedirectToAction(nameof(Index));
             }
+
+            ViewBag.Genders = EnumHelper.GetEnumSelectList<Gender>();
+            ViewBag.ContractTypes = EnumHelper.GetEnumSelectList<ContractType>();
+            ViewBag.Departments = EnumHelper.GetEnumSelectList<Department>();
+
             return View(employee);
         }
 
@@ -98,6 +112,11 @@ namespace CRUD_Employees.Controllers
             {
                 return NotFound();
             }
+
+            ViewBag.Genders = EnumHelper.GetEnumSelectList<Gender>();
+            ViewBag.ContractTypes = EnumHelper.GetEnumSelectList<ContractType>();
+            ViewBag.Departments = EnumHelper.GetEnumSelectList<Department>();
+
             return View(employee);
         }
 
@@ -117,9 +136,13 @@ namespace CRUD_Employees.Controllers
             {
                 try
                 {
-                    if (imageFile != null)
+                    var existingEmployee = await _context.Employees.FindAsync(id);
+                    if (existingEmployee == null)
                     {
-                        // Process the uploaded file
+                        return NotFound();
+                    }
+                    if (imageFile != null && imageFile.Length > 0)
+                    {
                         string uploadsFolder = Path.Combine(_webHostEnvironment.WebRootPath, "images");
                         string uniqueFileName = Guid.NewGuid().ToString() + "_" + imageFile.FileName;
                         string filePath = Path.Combine(uploadsFolder, uniqueFileName);
@@ -129,20 +152,36 @@ namespace CRUD_Employees.Controllers
                             await imageFile.CopyToAsync(fileStream);
                         }
 
-                        // Delete the old image if it exists from previous entry
-                        if (!string.IsNullOrEmpty(employee.Image))
+                        if (!string.IsNullOrEmpty(existingEmployee.Image) && existingEmployee.Image != "images/default_image.jpg")
                         {
-                            string oldFilePath = Path.Combine(_webHostEnvironment.WebRootPath, employee.Image);
+                            string oldFilePath = Path.Combine(_webHostEnvironment.WebRootPath, existingEmployee.Image);
                             if (System.IO.File.Exists(oldFilePath))
                             {
                                 System.IO.File.Delete(oldFilePath);
                             }
                         }
 
-                        // Store the new relative image path
-                        employee.Image = "images/" + uniqueFileName;
+                        existingEmployee.Image = "images/" + uniqueFileName;
                     }
-                    _context.Update(employee);
+
+                    if (employee.ContractType == ContractType.Permanent)
+                    {
+                        employee.ContractDue = null;
+                    }
+
+                    existingEmployee.FirstName = employee.FirstName;
+                    existingEmployee.LastName = employee.LastName;
+                    existingEmployee.Gender = employee.Gender;
+                    existingEmployee.BirthYear = employee.BirthYear;
+                    existingEmployee.StartedWorking = employee.StartedWorking;
+                    existingEmployee.ContractType = employee.ContractType;
+                    existingEmployee.ContractDue = employee.ContractDue;
+                    existingEmployee.Department = employee.Department;
+                    existingEmployee.VacationDays = employee.VacationDays;
+                    existingEmployee.DaysOff = employee.DaysOff;
+                    existingEmployee.PaidLeaveDays = employee.PaidLeaveDays;
+
+                    _context.Update(existingEmployee);                
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
@@ -158,6 +197,11 @@ namespace CRUD_Employees.Controllers
                 }
                 return RedirectToAction(nameof(Index));
             }
+
+            ViewBag.Genders = EnumHelper.GetEnumSelectList<Gender>();
+            ViewBag.ContractTypes = EnumHelper.GetEnumSelectList<ContractType>();
+            ViewBag.Departments = EnumHelper.GetEnumSelectList<Department>();
+
             return View(employee);
         }
 
@@ -185,24 +229,25 @@ namespace CRUD_Employees.Controllers
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
             var employee = await _context.Employees.FindAsync(id);
-            if (employee != null)
+            if (employee == null)
             {
-                // Delete the image file from wwwroot/images if it exists from previous entry
-                if (!string.IsNullOrEmpty(employee.Image))
+                return NotFound();
+            }           
+            if (!string.IsNullOrEmpty(employee.Image) && employee.Image != "images/default_image.jpg")
+            {
+                var imagePath = Path.Combine(_webHostEnvironment.WebRootPath, employee.Image);
+                if (System.IO.File.Exists(imagePath))
                 {
-                    var imagePath = Path.Combine(Directory.GetCurrentDirectory(), "wwwroot/images", employee.Image);
-                    if (System.IO.File.Exists(imagePath))
-                    {
-                        System.IO.File.Delete(imagePath);
-                    }
+                    System.IO.File.Delete(imagePath);
                 }
-
-                _context.Employees.Remove(employee);
             }
 
+            _context.Employees.Remove(employee);
             await _context.SaveChangesAsync();
+
             return RedirectToAction(nameof(Index));
         }
+
 
         private bool EmployeeExists(int id)
         {
